Tests/LibWeb/Text/expected/svg-a-element-activation-behavior.txt

basic link activated
target link activated in iframe
